move code files into sub-directory
- add octothorpe detection to logToHtml
